+++ src/Core/MarketHub.Application/DTOs/Persistence/Product/ProductDetailsDto.cs
@@ -15,7 +15,7 @@ public class ProductDetailsDto
     public string Type { get; set; } = string.Empty;
     public JsonElement Specifications { get; set; }
     public int NumberOfReviews { get; set; }
-    public int AverageRating { get; set; }
+    public decimal AverageRating { get; set; }
     public ICollection<ReviewDto>? Reviews { get; set; }
     public ICollection<string> ImagesUrls { get; set; } = new List<string>();
 }
